Moved epi_week logic to separate module meerkat_abacus.util.epi_week.
