+++ meerkat_abacus/data_management.py
@@ -17,7 +17,8 @@ from meerkat_abacus import model
 from meerkat_abacus import config
 from meerkat_abacus.codes import to_codes
 from meerkat_abacus import util
-from meerkat_abacus.util import create_fake_data, epi_week_for_date
+from meerkat_abacus.util import create_fake_data
+from meerkat_abacus.util.epi_week import epi_week_for_date
 import meerkat_libs as libs
 from shapely.geometry import shape, Polygon, MultiPolygon
 from geoalchemy2.shape import from_shape
